+++ js/components/firebase.factory.js
@@ -1,8 +1,8 @@
 // let's create a re-usable factory that generates the $firebaseAuth instance
 angular.module('cycloneApp').factory("firebaseRef", ['stateService',
-    function(stateService) {
+    function (stateService) {
 
-        function getTimeReference(user){
+        function getTimeReference(user) {
             // Load the current date (from stateService = URL) to load correct data
             currentDate = stateService.getCurrentDate();
             this.year = currentDate.year();
@@ -14,31 +14,31 @@ angular.module('cycloneApp').factory("firebaseRef", ['stateService',
             return reference;
         }
 
-        function getCurrentTaskReference(user){
+        function getCurrentTaskReference(user) {
             var ref = firebase.database().ref();
             var reference = ref.child("userSettings/" + user.uid + "/currentTask");
             return reference;
         }
 
-        function getProjectReference(user){
+        function getProjectReference(user) {
             var ref = firebase.database().ref();
             var reference = ref.child("userSettings/" + user.uid + "/projects");
             return reference;
         }
 
-        function getFeaturesReference(user){
+        function getFeaturesReference(user) {
             var ref = firebase.database().ref();
             var reference = ref.child("userSettings/" + user.uid + "/features");
             return reference;
         }
 
-        function getStartTimesReference(user){
+        function getStartTimesReference(user) {
             var ref = firebase.database().ref();
             var reference = ref.child("userSettings/" + user.uid + "/startTimes");
             return reference;
         }
 
-        function getWeeksOfMonthReferences(user){
+        function getWeeksOfMonthReferences(user) {
             const weekDate = stateService.getMonthDetails();
             const year = weekDate.year;
             const references = [];
@@ -47,31 +47,40 @@ angular.module('cycloneApp').factory("firebaseRef", ['stateService',
                 const ref = firebase.database().ref();
                 references.push(ref.child("time/" + user.uid + "/" + year + "/" + weekNumber));
             });
-            
+
             return {
                 details: weekDate,
                 references: references
             }
         }
 
+        function getTodoReference(user) {
+            var ref = firebase.database().ref();
+            var reference = ref.child("todo/" + user.uid);
+            return reference;
+        }
+
         return {
-            getTimeReference: function(user) {
+            getTimeReference: function (user) {
                 return getTimeReference(user);
             },
-            getCurrentTaskReference: function(user) {
+            getCurrentTaskReference: function (user) {
                 return getCurrentTaskReference(user);
             },
-            getProjectReference: function(user) {
+            getProjectReference: function (user) {
                 return getProjectReference(user);
             },
-            getFeaturesReference: function(user) {
+            getFeaturesReference: function (user) {
                 return getFeaturesReference(user);
             },
-            getStartTimesReference: function(user) {
+            getStartTimesReference: function (user) {
                 return getStartTimesReference(user);
             },
-            getWeeksOfMonthReferences: function(user) {
+            getWeeksOfMonthReferences: function (user) {
                 return getWeeksOfMonthReferences(user);
+            },
+            getTodoReference: function (user) {
+                return getTodoReference(user);
             },
         }
     }