+++ js/components/todo/todo-detail/todo-detail.component.js
@@ -1,0 +1,14 @@
+var todoDetail = {
+    bindings: {
+        todo: '<',
+        onSave: '&',
+        onUpdate: '&',
+        onDelete: '&'
+    },
+    template: require('./todo-detail.tpl.html'),
+    controller: 'TodoDetailController'
+};
+
+angular
+    .module('components.todo')
+    .component('todoDetail', todoDetail);