+++ js/components/todo/todo-new/todo-new.component.js
@@ -1,0 +1,20 @@
+var todoNew = {
+    template: require('./todo-new.tpl.html'),
+    controller: 'TodoNewController'
+};
+
+angular
+    .module('components.todo')
+    .component('todoNew', todoNew)
+    .config(function ($stateProvider) {
+        $stateProvider
+            .state('newTodo', {
+                parent: 'todo',
+                url: '/new',
+                views: {
+                    'head@todo': {
+                        component: 'todoNew'
+                    }
+                }
+            })
+    });